# optimize(Displayserver): performance up