Replace stdio with the posix api, and drop fscanf.

It shows a huge performance up in my local benchmarks.

About 35~40% faster per modefile than the old one.

Here is the highlights of benchmarks:

```text
Running ./bench_linux_amd64
Run on (8 X 1800.01 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB (x1)
Load Average: 0.12, 0.19, 0.16
-------------------------------------------------------
Benchmark             Time             CPU   Iterations
-------------------------------------------------------
BenchStdIO         1712 ns         1712 ns       341076
BenchPOSIXIO        998 ns          998 ns       655856
```

updates #634