@CarterLi 2023-11-28
Do you really think this PR can address any performance issues?

You are right that parsing string manually is always faster than fscanf, but why fscanf exists?

Come on! 1ms = 1000000ns. Several nanoseconds are not bottleneck, because it's executed only once. In addition, as for #634, this code is NOT executed unless `--ds-force-drm` is used.

Since @flipflop133 was using X11, the real bottleneck was the xrandr stuff. If you really want to improve fastfetch, please check that.