@CarterLi 2023-11-28
Don't get me wrong, I like performance improvements, and I'm accepting changes to improve performance at the cost of code complexity, but ONLY in common code.

Closing.