@CarterLi 2023-11-28
5x more lines of code and one more magic number are complexity.

> it even costs twice as much in performance

`1712ns -> 998ns` can be called twice.

`500 * 1000 * 1000ns + 1712ns -> 500 * 1000 * 1000ns + 998ns` can't.